Removed calls to legacy loaders from and models tests to individual model test classes from DataRepo.tests.models.test_peak_group_label.TracerRateTests.
Details:

- Corrected various imports of settings to import from django.conf.
- Changed a few tests to use different data to avoid repeated loads.

Files checked in: DataRepo/formats/dataformat_group.py DataRepo/tests/management/commands/test_load_table.py DataRepo/tests/models/test_infusate.py DataRepo/tests/models/test_peak_group.py DataRepo/tests/test_models.py DataRepo/tests/tracebase_test_case.py DataRepo/tests/views/upload/test_submission.py DataRepo/tests/models/test_animal_label.py DataRepo/tests/models/test_peak_group_label.py
