+++ DataRepo/tests/test_models.py
@@ -1578,333 +1578,6 @@ class PropertyTests(TracebaseTestCase):
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
-class MultiTracerLabelPropertyTests(TracebaseTestCase):
-    fixtures = ["data_types.yaml", "data_formats.yaml"]
-
-    @classmethod
-    @MaintainedModel.no_autoupdates()
-    def setUpTestData(cls):
-        call_command("loaddata", "lc_methods")
-        call_command(
-            "legacy_load_study",
-            "DataRepo/data/tests/small_obob/small_obob_study_prerequisites.yaml",
-        )
-        call_command(
-            "legacy_load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/data/tests/multiple_labels/animal_sample_table.xlsx"
-            ),
-            skip_researcher_check=True,
-        )
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/multiple_labels/alafasted_cor.xlsx",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-
-        super().setUpTestData()
-
-    def test_tracer_labeled_elements(self):
-        anml = Animal.objects.get(name="xzl1")
-        expected = ["C", "N"]
-        output = anml.infusate.tracer_labeled_elements
-        self.assertEqual(expected, output)
-
-    def test_serum_tracers_enrichment_fraction(self):
-        anml = Animal.objects.get(name="xzl5")
-        recs = anml.labels.all()
-        outputc = recs.get(element__exact="C").serum_tracers_enrichment_fraction
-        outputn = recs.get(element__exact="N").serum_tracers_enrichment_fraction
-        self.assertEqual(2, recs.count())
-        self.assertAlmostEqual(0.2235244143081364, outputc)
-        self.assertAlmostEqual(0.30075567022988536, outputn)
-
-    def test_peak_labeled_elements_one(self):
-        # succinate has no nitrogen
-        pg = PeakGroup.objects.filter(msrun_sample__sample__name="xzl5_panc").get(
-            name="succinate"
-        )
-        output = pg.peak_labeled_elements
-        # One common element
-        expected = ["C"]
-        self.assertEqual(expected, output)
-
-    def test_peak_labeled_elements_two(self):
-        pg = PeakGroup.objects.filter(msrun_sample__sample__name="xzl5_panc").get(
-            name="glutamine"
-        )
-        output = pg.peak_labeled_elements
-        expected = ["C", "N"]
-        self.assertEqual(expected, output)
-
-    def test_enrichment_abundance(self):
-        pg = PeakGroup.objects.filter(msrun_sample__sample__name="xzl5_panc").get(
-            name="glutamine"
-        )
-        pgc = pg.labels.get(element__exact="C").enrichment_abundance
-        pgn = pg.labels.get(element__exact="N").enrichment_abundance
-        expectedc = 1369911.2746615328
-        expectedn = 6571127.3714690255
-        self.assertEqual(pg.labels.count(), 2)
-        self.assertAlmostEqual(expectedc, pgc)
-        self.assertAlmostEqual(expectedn, pgn)
-
-    def test_normalized_labeling_2_elements(self):
-        pg = PeakGroup.objects.filter(msrun_sample__sample__name="xzl5_panc").get(
-            name="glutamine"
-        )
-        pgc = pg.labels.get(element__exact="C").normalized_labeling
-        pgn = pg.labels.get(element__exact="N").normalized_labeling
-        expectedc = 0.06287501342027346
-        expectedn = 0.2241489339907528
-        self.assertEqual(pg.labels.count(), 2)
-        self.assertAlmostEqual(expectedc, pgc)
-        self.assertAlmostEqual(expectedn, pgn)
-
-
-@override_settings(CACHES=settings.TEST_CACHES)
-class TracerRateTests(TracebaseTestCase):
-    @classmethod
-    def setUpTestData(cls):
-        call_command("loaddata", "lc_methods")
-        call_command(
-            "load_tissues",
-            infile="DataRepo/data/tests/small_obob2/tissues.tsv",
-        )
-        call_command(
-            "load_compounds",
-            infile="DataRepo/data/tests/small_obob2/compounds.tsv",
-        )
-
-        call_command(
-            "legacy_load_animals_and_samples",
-            sample_table_filename="DataRepo/data/tests/small_obob2/obob_samples_table.tsv",
-            animal_table_filename="DataRepo/data/tests/small_obob2/obob_animals_table.tsv",
-            table_headers="DataRepo/data/tests/small_obob2/sample_and_animal_tables_headers.yaml",
-        )
-
-        # for the fcirc and rate-calculation tests
-        call_command(
-            "legacy_load_accucor_msruns",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            accucor_file="DataRepo/data/tests/small_obob2/obob_maven_c160_serum.xlsx",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-        )
-
-        # defining a primary animal object for repeated tests
-        cls.MAIN_SERUM_ANIMAL = Animal.objects.get(name="970")
-
-        super().setUpTestData()
-
-    @tag("fcirc")
-    def test_peakgroup_is_tracer_label_compound_group(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        fpgl = pg.labels.filter(element=element).first()
-        self.assertTrue(fpgl.is_tracer_label_compound_group)
-        self.assertEqual(fpgl.peak_group.name, "C16:0")
-
-    @tag("fcirc")
-    def test_peakgroup_from_serum_sample(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        fpgl = pg.labels.filter(element=element).first()
-        self.assertTrue(fpgl.from_serum_sample)
-
-    @tag("fcirc")
-    def test_peakgroup_can_compute_tracer_label_rates(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        fpgl = pg.labels.filter(element=element).first()
-        self.assertTrue(fpgl.can_compute_tracer_label_rates)
-
-    @tag("fcirc")
-    def test_peakgroup_can_compute_body_weight_intact_tracer_label_rates(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        fpgl = pg.labels.filter(element=element).first()
-        self.assertTrue(fpgl.can_compute_body_weight_intact_tracer_label_rates)
-
-    @tag("fcirc")
-    def test_peakgroup_can_compute_body_weight_average_tracer_label_rates(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        fpgl = pg.labels.filter(element=element).first()
-        self.assertTrue(fpgl.can_compute_body_weight_average_tracer_label_rates)
-
-    @tag("fcirc")
-    def test_peakgroup_can_compute_intact_tracer_label_rates(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        fpgl = pg.labels.filter(element=element).first()
-        self.assertTrue(fpgl.can_compute_intact_tracer_label_rates)
-
-    @tag("fcirc")
-    def test_peakgroup_can_compute_average_tracer_label_rates(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        fpgl = pg.labels.filter(element=element).first()
-        self.assertTrue(fpgl.can_compute_average_tracer_label_rates)
-
-    @tag("fcirc")
-    def test_nontracer_peakgroup_calculation_attempts(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        nontracer_compound = Compound.objects.get(name="succinate")
-        non_tracer_pg_label = (
-            animal.last_serum_sample.msrun_samples.first()
-            .peak_groups.get(compounds__exact=nontracer_compound)
-            .labels.first()
-        )
-        # # but let's get a peakgroup for a compound we know is not the tracer
-        # pgs = animal.last_serum_sample.peak_groups(nontracer_compound)
-        # # should only be one in this specific case
-        # non_tracer_pg = pgs[0]
-        # tryptophan is not the tracer
-        self.assertFalse(non_tracer_pg_label.is_tracer_label_compound_group)
-        # and none of these should return a value
-        with self.assertWarns(UserWarning):
-            self.assertIsNone(non_tracer_pg_label.rate_disappearance_intact_per_gram)
-        with self.assertWarns(UserWarning):
-            self.assertIsNone(non_tracer_pg_label.rate_appearance_intact_per_gram)
-        with self.assertWarns(UserWarning):
-            self.assertIsNone(non_tracer_pg_label.rate_disappearance_intact_per_animal)
-        with self.assertWarns(UserWarning):
-            self.assertIsNone(non_tracer_pg_label.rate_appearance_intact_per_animal)
-        with self.assertWarns(UserWarning):
-            self.assertIsNone(non_tracer_pg_label.rate_disappearance_average_per_gram)
-        with self.assertWarns(UserWarning):
-            self.assertIsNone(non_tracer_pg_label.rate_appearance_average_per_gram)
-        with self.assertWarns(UserWarning):
-            self.assertIsNone(non_tracer_pg_label.rate_disappearance_average_per_animal)
-        with self.assertWarns(UserWarning):
-            self.assertIsNone(non_tracer_pg_label.rate_appearance_average_per_animal)
-
-    @tag("fcirc")
-    def test_last_serum_tracer_rate_disappearance_intact_per_gram(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        pgl = pg.labels.filter(element=element).first()
-        self.assertAlmostEqual(
-            pgl.rate_disappearance_intact_per_gram,
-            38.83966501,
-            places=2,
-        )
-
-    @tag("fcirc")
-    def test_last_serum_tracer_rate_appearance_intact_per_gram(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        pgl = pg.labels.filter(element=element).first()
-        self.assertAlmostEqual(
-            pgl.rate_appearance_intact_per_gram,
-            34.35966501,
-            places=2,
-        )
-
-    @tag("fcirc")
-    def test_last_serum_tracer_rate_disappearance_intact_per_animal(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        pgl = pg.labels.filter(element=element).first()
-        self.assertAlmostEqual(
-            pgl.rate_disappearance_intact_per_animal,
-            1040.903022,
-            places=2,
-        )
-
-    @tag("fcirc")
-    def test_last_serum_tracer_rate_appearance_intact_per_animal(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        pgl = pg.labels.filter(element=element).first()
-        self.assertAlmostEqual(
-            pgl.rate_appearance_intact_per_animal,
-            920.8390222,
-            places=2,
-        )
-
-    @tag("fcirc")
-    def test_last_serum_tracer_rate_disappearance_average_per_gram(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        pgl = pg.labels.filter(element=element).first()
-        self.assertAlmostEqual(
-            pgl.rate_disappearance_average_per_gram,
-            37.36671487,
-            places=2,
-        )
-
-    @tag("fcirc")
-    def test_last_serum_tracer_rate_appearance_average_per_gram(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        pgl = pg.labels.filter(element=element).first()
-        self.assertAlmostEqual(
-            pgl.rate_appearance_average_per_gram,
-            32.88671487,
-            places=2,
-        )
-
-    @tag("fcirc")
-    def test_last_serum_tracer_rate_disappearance_average_per_animal(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        tracer = animal.infusate.tracers.first()
-        element = tracer.labels.first().element
-        pg = animal.last_serum_tracer_peak_groups.first()
-        pgl = pg.labels.filter(element=element).first()
-        # doublecheck weight, because test is not exact but test_tracer_Rd_avg_g was fine
-        self.assertAlmostEqual(
-            pgl.rate_disappearance_average_per_animal,
-            1001.427958,
-            places=2,
-        )
-
-    @tag("fcirc")
-    def test_last_serum_tracer_rate_appearance_average_per_animal(self):
-        animal = self.MAIN_SERUM_ANIMAL
-        # Uses Animal.last_serum_sample and Sample.peak_groups
-        pgl = animal.last_serum_sample.last_tracer_peak_groups.first().labels.first()
-        self.assertAlmostEqual(
-            pgl.rate_appearance_average_per_animal,
-            881.3639585,
-            places=2,
-        )
-
-
-@override_settings(CACHES=settings.TEST_CACHES)
 class AnimalAndSampleLoadingTests(TracebaseTestCase):
     @classmethod
     def setUpTestData(cls):
